Restructuring of Code for Creating new Process
It has some changes in structure for putting the test cases
dependencies under unit-test.
