+++ core/decorators/worker.ts
@@ -1,33 +1,35 @@
 require("reflect-metadata/Reflect");
 import * as express from "express";
-import * as Config from "./config";
-import * as securityConfig from "./security-config";
-import {router} from "./core/exports";
-import * as data from "./mongoose";
-import { Decorators } from './core/constants/decorators';
-import { DecoratorType } from './core/enums/decorator-type';
-import { MetaUtils } from "./core/metadata/utils";
-import {MetaData, IMetaOptions} from "./core/metadata/metadata" ;
-import {CurrentUserDetailService} from "./current-user-detail-service";
-import {UserDetailService} from "./security/auth/user-detail-service";
+import * as Config from "../../config";
+import * as securityConfig from "../../security-config";
+import {router} from "../exports";
+import * as data from "../../mongoose";
+import { Decorators } from '../constants/decorators';
+import { DecoratorType } from '../enums/decorator-type';
+import { MetaUtils } from "../metadata/utils";
+import {MetaData, IMetaOptions} from "../metadata/metadata" ;
+import {CurrentUserDetailService} from "../../current-user-detail-service";
+import {UserDetailService} from "../../security/auth/user-detail-service";
 import * as Enumerable from 'linq';
-import * as Utils from "./core/utils";
-import {winstonLog} from './logging/winstonLog';
-import {Container} from './di';
+import * as Utils from "../utils";
+import {winstonLog} from '../../logging/winstonLog';
+import {Container} from '../../di';
 
 var serverUp= function(){
     winstonLog.logInfo("+++++\n\n\n\++++++++++++++ ======= Executed +++\n\n\n\n\++++++++++++++++");
     const app = express();
-    const Main = require("./core");
+    const Main = require("../../core");
     Config.Config.ignorePaths= Config.Config.ignorePaths || [];
     Config.Config.ignorePaths.push('**/server.js','**/worker.js')
     Main(Config, securityConfig, __dirname, data.entityServiceInst);
+    const test = require('../../unit-test/services/blogServiceImpl'); // Test Service i.e. BlogService required for testing in Jasmine.
     data.connect();
     data.generateSchema();
     app.use("/", router);
     Main.register(app);
 }
 process.on('message', function (m) {
+    try{
     serverUp(); // Prepares the server.
     MetaUtils.childProcessId = process.pid;
     winstonLog.logInfo("Input from Parent Process : "+JSON.stringify(m));
@@ -35,7 +37,7 @@ process.on('message', function (m) {
     if( m.worker_params != null) {
         var services = MetaUtils.getMetaDataForDecorators([Decorators.SERVICE]);
         var service_name=m.worker_params.serviceName;
-        winstonLog.logDebug(" All Available Services : "+ JSON.stringify(services) + " service: "+ service_name);
+        //winstonLog.logDebug(" All Available Services : "+ JSON.stringify(services) + " service: "+ service_name);
         var service=Enumerable.from(services).where(x => x.metadata[0].target.constructor.name == service_name).select(x => x.metadata[0]).firstOrDefault();
         winstonLog.logInfo("Required Service: "+ service + " with input content : "+ JSON.stringify(service));
         if(service){
@@ -63,4 +65,7 @@ process.on('message', function (m) {
      }else{
          process.send("Worker param is null. Exiting.... ");
      }
+    }catch(error){
+        winstonLog.logError("Error found: "+ error);
+    }
  });
